requirements.txt
flask
pillow
pyzbar
qrcode
gunicorn